+++ src/MailerSend.php
@@ -3,9 +3,12 @@
 namespace MailerSend;
 
 use MailerSend\Common\HttpLayer;
+use MailerSend\Endpoints\Activity;
+use MailerSend\Endpoints\Analytics;
 use MailerSend\Endpoints\Domain;
 use MailerSend\Endpoints\Email;
 use MailerSend\Endpoints\Message;
+use MailerSend\Endpoints\Webhook;
 use MailerSend\Endpoints\Token;
 use MailerSend\Exceptions\MailerSendException;
 use Tightenco\Collect\Support\Arr;
@@ -31,7 +34,10 @@ class MailerSend
 
     public Email $email;
     public Message $messages;
+    public Webhook $webhooks;
     public Token $token;
+    public Activity $activity;
+    public Analytics $analytics;
     public Domain $domain;
 
     /**
@@ -50,7 +56,10 @@ class MailerSend
     {
         $this->email = new Email($this->httpLayer, $this->options);
         $this->messages = new Message($this->httpLayer, $this->options);
+        $this->webhooks = new Webhook($this->httpLayer, $this->options);
         $this->token = new Token($this->httpLayer, $this->options);
+        $this->activity = new Activity($this->httpLayer, $this->options);
+        $this->analytics = new Analytics($this->httpLayer, $this->options);
         $this->domain = new Domain($this->httpLayer, $this->options);
     }